requirements-dev.txt
# Include production requirements
-r requirements.txt

# Development Tools
ipython==8.18.1
ipdb==0.13.13

# Testing Tools  
pytest-mock==3.12.0
pytest-env==1.1.3
httpx-mock==0.4.0
faker==20.1.0

# Code Quality Tools
pre-commit==3.6.0
isort==5.13.2
autopep8==2.0.4
pylint==3.0.3

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings[python]==0.24.0

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Security Scanning
bandit==1.7.6
safety==3.0.1